apps/sandbox_workflows/pyproject.toml
[project]
name = "obox"
version = "0.1.0"
description = "Orchestrated Sandbox Workflows - Multi-agent experimentation with E2B sandboxes"
requires-python = ">=3.12"
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "claude-agent-sdk>=0.1.0",
    "python-dotenv>=1.0.0",
    "e2b>=2.6.4",
]

[project.scripts]
obox = "src.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
